move envers config
